src/vanilla/requirements.txt
revChatGPT
python-dotenv